+++ projects/igniteui-angular/src/lib/core/i18n/grid-resources.ts
@@ -66,6 +66,8 @@ export interface IGridResourceStrings {
     igx_grid_excel_apply?: string;
     igx_grid_excel_search_placeholder?: string;
     igx_grid_excel_select_all?: string;
+    igx_grid_excel_select_all_search_results?: string;
+    igx_grid_excel_add_to_filter?: string;
     igx_grid_excel_blanks?: string;
     igx_grid_excel_hide?: string;
     igx_grid_excel_show?: string;
@@ -178,6 +180,8 @@ export const GridResourceStringsEN: IGridResourceStrings = {
     igx_grid_excel_apply: 'apply',
     igx_grid_excel_search_placeholder: 'Search',
     igx_grid_excel_select_all: 'Select All',
+    igx_grid_excel_select_all_search_results: 'Select all search results',
+    igx_grid_excel_add_to_filter: 'Add current selection to filter',
     igx_grid_excel_blanks: '(Blanks)',
     igx_grid_excel_hide: 'Hide column',
     igx_grid_excel_show: 'Show column',