+++ projects/igniteui-angular/src/lib/grids/row.directive.ts
@@ -392,7 +392,7 @@ export class IgxRowDirective<T extends IgxGridBaseDirective & GridType> implemen
      */
     public update(value: any) {
         const crudService = this.crudService;
-        if (crudService.inEditMode && crudService.cell.id.rowID === this.rowID) {
+        if (crudService.cellInEditMode && crudService.cell.id.rowID === this.rowID) {
             this.grid.endEdit(false);
         }
         const row = new IgxRow(this.rowID, this.index, this.rowData, this.grid);